snap/snapcraft.yaml: Build libvlc from its sources
Instead of keeping the old libvlc found on Ubuntu, use the
VLC version 3.0 source, in order to work with video streams
with uses H.264/H.265.

Building libvlc on snap is not an easy task, as it uses
Ubuntu xenial repositories, with contains a ffmpeg library
that it is too old for VLC 3.0. That adds an extra
complexity to libVLC build.

Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>