README.md: update build instructions for Fedora
While playing with Docker in order to try building a kaffeine
container, I noticed that the Fedora build instructions are incomplete,
as they miss gcc/cmake/make. Add it and do a cleanup to make it easier
for its usage.

Also, as since Fedora 23 the instructions are the same, remove the
note with mentions on what versions it is known.

Tested with Fedora 28.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>